+++ src/cx_mesh_gen.h
@@ -1,6 +1,8 @@
 #ifndef CX_MESH_GEN_H
 #define CX_MESH_GEN_H
 
+#include <stdint.h>
+
 struct he_mesh;
 struct mesh_primitive;
 
@@ -14,17 +16,17 @@ void cx_mesh_gen_box(const float p_hald_size[3], struct mesh_primitive* p_out_me
 
 void cx_mesh_gen_sphere(
 	const float radius,
-	const unsigned int rings, const unsigned int segments,
+	const uint16_t rings, const uint16_t segments,
 	struct mesh_primitive* p_out_mesh_primitive);
 
 void cx_mesh_gen_hemisphere(
 	const float radius,
-	const unsigned int rings, const unsigned int segments,
+	const uint16_t rings, const uint16_t segments,
 	struct mesh_primitive* p_out_mesh_primitive);
 
 void cx_mesh_gen_cylinder(
 	const float radius_a, const float radius_b, const float half_length,
-	const unsigned int segments,
+	const uint16_t segments,
 	const int b_cap_a, const int b_cap_b,
 	struct mesh_primitive* p_out_mesh_primitive);
